fix(gateway): Enforce 200 OK for Place Order API
Strictly return HTTP 200 OK for order placement to meet API contract requirements, removing the previous 202 Accepted handling.

+++ src/internal_transfer/adapters/funding.rs
@@ -399,7 +399,7 @@ impl ServiceAdapter for FundingAdapter {
             WHERE user_id = $2 AND asset_id = $3 AND account_type = $4
             "#,
         )
-        .bind(amount as i64) // amount is i64 here from rollback query
+        .bind(amount) // amount is i64 here from rollback query
         .bind(user_id)
         .bind(asset_id)
         .bind(FUNDING_ACCOUNT_TYPE)
